requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
pandas==2.1.3
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
faiss-cpu==1.13.1
google-generativeai==0.3.1
python-dotenv==1.0.0
pydantic==2.5.0
numpy==1.26.2
tqdm==4.66.1
selenium==4.15.2
webdriver-manager==4.0.1
sentence-transformers==2.2.2
torch

